backend/requirements.txt
Flask
mercadopago
python-dotenv
flask-cors
gunicorn
psycopg2-binary
